requirements.txt
Pillow
rembg
onnxruntime
pyinstaller
customtkinter
cairosvg
